requirements.txt
cuvis==3.3.0.post1
lightly==1.5.19
lightning==2.5.0.post0
matplotlib==3.10.0
numpy==1.26.4
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
pytorch_lightning==2.5.0.post0
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
torch==2.6.0+cu124
torchmetrics==1.6.1
torchvision==0.21.0+cu124
tqdm==4.67.1
tensorboard==2.19.0